elfeed: Update feed URLs to canonical and HTTPS
Sometimes feeds change location over time. Fortunately most websites
handle this pretty well and redirect from the old location forever.
However, it would be much better to never need to do this in the first
place! Especially if the feed moves an entirely new domain. It's
technically a new feed at that point, and entries can have identity
issues (more of a problem for RSS than Atom).

There's no way to future-proof the protocol part of a feed URL. We can't
anticipate what protocols we'll be using in the future and what they'll
be called. The world is a different place than it was when many of these
feeds were first created. Technology and circumstances have changed.

For feeds that are served properly over HTTPS, I've switched to HTTPS.
This saves a redirect in cases where HTTPS is now forced. However, many
sites:

1. Don't accept connections on port 443
2. Respond with the wrong protocol on port 443
3. Serve long-expired certificates
4. Serve the wrong certificate (e.g. any Tumblr site with a custom domain)
5. Serve a self-signed certificate (tedu!)
6. Redirect back to plain HTTP (e.g. any Tumblr site, which also adds
   fragment garbage when redirecting).

Now, I'm no HTTPS zealot. Like nearly every technology, there are
trade-offs between HTTPS and HTTP. HTTPS isn't always better. For
example, after this change I expect the number of fetch failures to
increase because there is a new class of new error condition that
doesn't exist with plain HTTP: certificate errors. I will run into
expired certificates on occasion, or invalid certificates when someone
reconfigures their host too hastily. In return for accepting an
increased failure rate, I get a theoretical increase in security that,
for my needs, serves no practical purpose.

It's probably a bit slower, too, due to the extra handshaking and the
cryptography overhead. Though I doubt it's enough to be noticeable in
Elfeed. Using TLS potentially grants access to HTTP/2, which is
generally more efficient than HTTP/1.1. However, the efficiencies of
HTTP/2 don't apply to fetching a single resource like a syndication
feed. (Conversely, if misconfigured, like, say, to server-push
unnecessary resources, HTTP/2 can actually be worse!)

So, in the case of syndication feeds, I generally don't believe it's a
moral imperative for hosts to switch over to HTTPS. It's not necessarily
worth the cost. That being said, of the six situations I listed, only 1
and 5 are acceptable host configurations. Either fully commit to using
HTTPS correctly or don't use it at all. No half measures. Item 5 is
irritating and wrong from a pragmatic standpoint, but I accept it as a
form of protest against the ridiculous CA situation (despite the
existence of Let's Encrypt).